tests: add a utility for property-based tests around an equivalence
This change introduces a utility type which was previously proposed for
the quickcheck library. The utility is not (yet) part of quickcheck
itself, but we want to use it now. Luckily, quickcheck's license allows
copying parts of the code, and I'm the original author of the type,
anyway. We therefore introduce the type locally, for now.

Hopefully, it will get accepted and released as part of quickcheck
itself, eventually.

fixup! test: add a utility for property-based tests around an equivalence