parsers: add module for parser utilities
We'll place parsers for FIRRTL constructs in the modules corresponding
to those constructs. I.e. the parser will be constructed from
sub-parsers distrubuted among modules. However, there will likely be
some common constructs used in multiple parsers in different modules.

This change introduces a module for those constructs with a result and
an error type to use at least for internal parsers.