types: add types explicitly for bit-widths and vector-widths
These are semantically different things. And bit-widths may be
undefined. In addition, both widths can never be zero, which we did not
model properly. In the case of a bit-width, we could have used zero to
represent an undefined value, but the introduced `Option` is type-safe
and will probably boil down to the same thanks to some optimization.