parsers: add function for convenient consumption of spaces
Previously, we would have our base parsers simply consume all preceding
spaces. However, this turned out to be an ill choice as FIRRTL is one of
those languages where the indentation has meaning. We therefore decided
to move all that whitespace consumption out of the base parsers.

Still, we'll want to avoid putting `space0`s everywhere, either in
conjunction with a `preceded` or, even worse, throwing our the result in
a `map`. Hence, we decided to introduce a convenience function which
hopefully helps avoid unnecessary boilerplate, as a prequisite for the
migration.