types: migrate type parsers to explicit handling of spaces
Previously, we would have our base parsers simply consume all preceding
spaces. However, this turned out to be an ill choice as FIRRTL is one of
those languages where the indentation has meaning. We therefore decided
to move all that whitespace consumption out of the base parsers.

This change migrates the parsers for both ground types and complex
types. As a result, the parsers no longer consume preceding whitespace.