indentation: create a type representing indentation
FIRRTL is one of those languages where indentation is syntactically
relevant. Therefore, we need to make sure we follow the indentation
rules when printing FIRRTL and we also need to recognize indentation
when parsing.

This change introduces a new module. In the future, it will provide
utilities for both printing and parsing indentation. This change also
introduces a type for representing an indentation level.