lib: expose Circuit parser
We recently introduced parsers for various elements which build upon
another, concluding with a parser for `Circuit`s. The general motivaiton
was, of course, to provide parsing capabilities to users of the library.
However, we consider most of the parsers internal and we chose not to
expose our internal module structure. Hence, all those parsers were, in
fact, hidden from users.

The one parser which we absolutely do want to expose is the
aforementioned parser for `Circuit`s. This change does so via a
re-export. Future changes may introduce some additional convenience
functions, e.g. for parsing a `Circuit` from a file.