types: add type for individual bundle fields
Previously, we would use a simple tuple for representing a bundle field.
This resulted in some constructs in `Type` which are a bit unwieldy.
This change introduces a dedicated struct for those fields. Future
changes will migrate `Type` to this new type. `OrientedType`, however,
will still use a simple tuple for bundle fields.