types: repurpose TypeEq trait for common type concepts
Previously, the `TypeEq` was used exclusively for providing a common
interface for querying type equivalence. However, there are also othe
concepts such as passiveness and inferencing, which we'll want to
express for at least `GroundType` and `Type`, and which would also be
applicable to `OrientedType`.

Rather than introducing new traits, we decided to repurpose the `TypeEq`
trait. This change does so by simply renaming the trait and moving
around some documentaiton.