module: remove member module from Port struct
Initially, we assumed that we'd need to refer to ports from outside a
module, eventually. We would have refered to such a port through an
`Arc<Port>` alone, and have the port refer to its parent module.
However, references to a port turned out to be valid only from inside
the module. Instantiations of a module effectively introduce a bundle
representing an interface, with port types potentially deviating due to
inferencing.

Since we'll refer to a given `Port` only inside the module, the
information which module the port belongs to is no longer required.