types: add trait for typed entities
We usually assign types to entities of some sort. For example, FIRRTL
expressions and module ports are typed. This change introduces a trait
for these entities, thus providing a common interface to query the type
of all typed entities. Future changes will implement this trait for
various types of entities.