expr: impl Typed for Expression
Naturally, FIRRTL expressions are typed. However, them being typed
requires the underlying references to be typed. This is not the case for
the entities we use for testing the parser, and having those tests
satisfy typing rules would complicate things unnecessarily.

Instread, we express this dependency directly by implementing `Typed`
only if the references are also `Typed`.