types: impl From<Vec<BundleField>> for Type
Constructing a `Type::Bundle` from a sequence of `BundleField`s is
something we suspect will turn up a few times. The variant does consist
of an `Arc` over a slice, but in practive we'll likely use a `Vec` for
collecting such items, anyway. Thus, a conversion from `Vec<BundeField>`
should be the most useful.