types: add trait for combining two entities
For an implementation of type inference, but also when determining the
type of a binary operation (e.g. an addition), we need to merge two
stucturally similar FIRRTL types.

This change introduces a trait providing an interface for performing
such a merge or combination. Future changes will implement this trait
for `Type` and `GroundType`.