expr: add functions for constructing some type variants
During testing, we generate arbitrary expressions. Currently, we do
not respect typing constraints. Thus, we may generate expressions which
are invalid, e.g. a sub-index on a UInt-literal. We plan on changing
this, i.e. generating expressions which could at least be typed
correctly.

During generation, we naturally need to generate sub-expressions. Their
types are related to the overall expressions, i.e. we need to derive the
sub-expression's type from the type of the current node. The functions
introduced with this chage function as helpers for this purpose.