expr: add functions for expression generation
During testing, we generate arbitrary expressions. Currently, we do
not respect typing constraints. Thus, we may generate expressions which
are invalid, e.g. a sub-index on a UInt-literal. We plan on changing
this, i.e. generating expressions which could at least be typed
correctly.

This change introduces functions generating expressions based on a given
type.