expr: add helper for generating typed expressions
We want to test the `Typed` impl for `Expression`. The most simple test
would just query the type, making sure it doesn't produce an error (all
the expressions we generate are supposed to be well-typed).

However, since we generate expressions from a given type using the
recently introduced fn `expr_with_type`, we'd also want to compare the
type returned by `Typed::type` with the one we used for generating the
expression.

With `quickcheck` the generation is done outside the test function, and
we don't have access to the RNG inside the funciton. We could just
create a new one, but that would be not ideal. Instead, we define a
helper in the form of a struct which encapsulates a type and an
expression generated from that type. The struct implements `Arbitrary`,
which allows generating expressions with a known target type as well as
(to some extend) shrinking preserving that relationship.