types: add fn for generating a map containing BundleFields
During testing, we previously generated a bundle by generating
`BundleFields` individually and collecting them in a `Vec`. Thus, we
could generate bundles with multiple fields of the same name. We decided
to work around this issue by inserting each field into a map which maps
a name to a field with that name.

This change introduces a function for generating such a map. In the
future, we may want to use this function in contexts where we insert
specially crafted fields. Thus, the function is public and returns the
map rather than a readily generated type.