types: use bundle_fields in Arbitrary impl for Type
During testing, we previously generated a bundle by generating
`BundleFields` individually and collecting them in a `Vec`. Thus, we
could generate bundles with multiple fields of the same name. We decided
to work around this issue by inserting each field into a map which maps
a name to a field with that name.

This change implements this approach using the recently introduced
`bundle_fields` function.