types: remove test for passiveness preservation for oriented type
We assumed that the passiveness property would be conserved when
converting a type to an oriented type. However, at least one corner-case
exists where this is not the case.

The following two types result in the same oriented type:

    {flip a: {flip a: UInt}}
    {a: {a: UInt}}

While the former is not passive, the latter is, and so is the oriented
type corresponding to both.