expr: remove parse check for primitive operations
Parsing of primitive operations are already covered by the `parse_expr`
test. It does have a higher probability of hitting interesting cases
for primitive operations, but the primary reason we introduced this test
in the first place was to get more usable reports.

Due to improvements in our shrinkers, this is hopefully not necessary
any more.