module: use an `Arc<str>` for storing identifier of a `Port`
When introducing `Port`, we assumed that we would not need to refer to
that identifier anywhere else. However, the type of a module instance,
i.e. an instantiation, is defined as a bundle with one field for each of
the module's ports. The field will share the associated port's name.

Using an `Arc` rather than a `String` allows sharing the memory
allocated for the string data with little to no additional cost.