types: use builder-function for controlling a BundleFields orientation
When constructing passive types, needing to supply the orientation each
single time is rather anoying. Hence, we decided to drop that parameter
from `BundleFields::new` and use a builder-like function instead. For
more convenience, we also supply a similar function which flips the
orientaiton. Such a function should come in handy both when implementing
some transformations and when construcing bundles by hand.