types: restrict publicness of parsers submodule to crate
In the past, we choose to keep this crate's inner structure hidden from
users and re-export all types and functions we actually want to provide
from the toplevel module. However, while this restrictive approach
(presumably) prevented us from leaking inner details, it imposed the
necessity of selecting each and every type a user may need to construct
in order to use the library. In addition, it created the problem of
potential naming conflicts, such as for `module::Port` and
`memory::Port`.

We decided to make most modules public instead. Still, many of those
will contain entities which need to be public to other modules, but
which we want to keep hidden from the outside, e.g. most parsers and
parsing utilities. Fortunately, rust allows to make a module public only
to other modules in the crate via `pub(crate)`.

This change does so for `types::parsers`.