lib: make circuit, expr, module and types public modules
In the past, we choose to keep this crate's inner structure hidden from
users and re-export all types and functions we actually want to provide
from the toplevel module. However, while this restrictive approach
(presumably) prevented us from leaking inner details, it imposed the
necessity of selecting each and every type a user may need to construct
in order to use the library. In addition, it created the problem of
potential naming conflicts, such as for `module::Port` and
`memory::Port`.

We decided to make most modules public instead.