lib: reduce number of types re-exported from lib
In the past, we choose to keep this crate's inner structure hidden from
users and re-export all types and functions we actually want to provide
from the toplevel module. However, while this restrictive approach
(presumably) prevented us from leaking inner details, it imposed the
necessity of selecting each and every type a user may need to construct
in order to use the library. In addition, it created the problem of
potential naming conflicts, such as for `module::Port` and
`memory::Port`.

We therefore made modules which contain types we do want to export
public. Hence, we can reduce the number of re-exports to types which are
of primary interest.