stmt: add a data type representing a referencable entity
For references in expressions, we defined the `expr::Reference` trait.
However, up until now we did not have a type we could have used for
actual FIRRTL expressions. Rather, we implemented `expr::Reference` for
types we'd use in tests and specific entities such as module ports and
registers.

This change introduces a type covering all the entities which may appear
in a FIRRTL expression in a new module which will define data types for
FIRRTL statements in the future.

Ultimately this will result in a circular dependency between `stmt` and
`module`, as a module will usually contain statements. The alternative
would be to have `Entity` defined in `module` and ride the generic
`Reference` train a little longer. However, this would make parsers a
bit awkward.