indentation: introduce a parser for indentation taking ownership
We already do have a parser for parsing indentations. The existing
parser borrows an indentation mutually, leaving it in a locked state
after a successful parse. This interface proved to be a good choice for
structures which do not involve alternatives at the top-level. However,
if branching parsers are involved, the approach does not work since it
would require multiple mut borrows.

For those cases, we introduce a second parser which owns an indentation
and yields a copy of the updated indentation after the parse. This
allows using the parser in branching parsers. However, it also requires
the user to update the indentation from the result.