display: widen contract for CommaSeparator to allow IntoIterator + Clone
The recently introduced `CommaSeparator` turned out to be not as useful
as we thought it would be, as it did not allow filtering and/or mapping
items prior to printing. Collecting them before instantiation was
possible, but would introduce unnecessary overhead.

Hence, we decided to make it accept types which implement `IntoIterator`
with a displayable item type. Since `Display` mandates that we do not
mutate the struct, we further require that type to implement `Clone`.