stmt: add parsers for statement lists and individual statements
We introduce both together since they depend on another. We also
introduce a private `indented_condition` as an abstraction.

Contrary to the "top-level" parsers in this module, it receives a
dynamically dispatched reference for the `Reference` factory rather than
an `impl`. The reason for this is the need for type-erasure to avoid
unbounded type recursion. This issue arises for the `Reference` factory
because we incorporate previously parsed declarations in `stmts`.