parsers: add utility parser for spaced line endings
Previously, we'd use a bare `nom::character:complete:::line_ending`
parser for accepting the end of a line. In virtually all cases, we'd
not even accept whitespace preceding the line ending.

In order to improve the situation, we decided to use a convenience
parser which does accept (and consume) preceding whitespace. In
addition, we also make it consume any empty/whitespace-only lines which
may occur in FIRRTL code.

This change introduces such a parser. Future changes will migrate
parsers for various constructs from bare `line_ending`s to this new
utility parser.