expr: add convenience fns for generating sink or source flows for tests
We plan on considering the expected flow of an expression in addition to
its type. In some situations we need to supply a flow which is suitable
for either a sink or source. Specifying `Flow::Sink` or `Flow::Source`
would be too restrictive since `Flow::Duplex` may also be a valid choice
in both cases.