circuit: make circuit store only the top module
When introducing the type initially, our understanding of module
instantiation FIRRTL was still incomplete. This would lead to a design
which would have required the top-level circuit to keep `Module`s
afloat. However, as instantiations have ownership of an `Arc`
corresponding to a module, the only module we need to keep alive in the
circuit is the top-level one.

Thus, we can remove that data dependency. This will ultimately make the
printing code more complex, but that's not really a problem.