module: differentiate between Module Kinds
The FIRRTL spec describes two distinct types of modules: `module` and
`extmodule`. Up until now, our `Module` type only covered `module`s.
Without statements.

This change allows differentiating between external and "regular"
modules. We don't use the `Kind` type internally since it was
introduced for user-faced differentiation. Using it internally would
require us to manually maintain consistency as we would need to store
statements for regular modules while making sure that external modules
do _not_ have statements,

Rather, we make `Module` carry either a list of statements or an
indication that it doesn't have the ability to do so. In the latter
case, the module must be an external one, which allows us to derive the
`Kind` of a module.