module: make module parser parse statements
Regular modules will contain statements, which we need to parse. This
involves resolving references, which is already done inside the `stmts`
parser. However, this also includes referencing modules, by name, for
instantiaitons. Thus the `module` parser gains a new parameter for
module resolution, which needs to be passed on call-sites.

For the `circuit` parser, we supply a dummy resolution which will not
work for modules with actual dependencies. A future changes will resolve
this issue.