module: add utility for parsing dependent modules
Parsing a module involves parsing statements, including potential
declarations of module instances, for which we need to resolve module
names.

Modules (usually) appear in lists, with a module appearing only after
its dependencies. In order to make parsing such lists of modules more
convenient, we introduce this utility, which functions as an `Iterator`
designed to interact nicely with a `circuit::ModuleConsumer`.