circuit: recursively print all dependencies of the top module
In the past, a circuit would be completely defined by its top module.
However, we recently allowed `Module`s to carry statements, including
instantiations of modules.

As a circuit is meant to be self-sufficient, we need to include all the
dependencies of a formatted module, i.e. all modules instantiated in it.