stmt: sort reference and modules by name for parse_stmt test
When writing the test initially, we were a bit lazy. We'd collect
referenced entities and modules in `Vec`s and collect their names into a
set in order to check uniqueness. We wouldn't even use the benefits of a
hash map for lookup, resulting in `O(n^2)` complexity for the overall
test.

Just sorting the items in those `Vec`s by name allows both efficient
checks for uniqueness without allocating additional memory and use a
banary search for the lookup.