stmt: make Statement a struct
When introducing `Statement`s, we primarily had to accomodate a set of
statement types with little to no shared structure. Hence, we choose to
make it an `enum` with one variant for each type of statement --with
declarations being collected into a single type.

However, any FIRRTL statement may be annotated with an "info" attribute,
which we plan on introducing as a property of `Statement` in the future.
We want to avoid duplicating such inforamtion in every single type of
statement as it would complicate virtually everything, including
accessing the attribute. Ideally, we'd like a struct to which we simply
add a field.

This change makes `Statement` a struct which may be extended with such a
field by moving the definition of statement types into a new enum
`Kind`. This naturally causes a lot of changes in `stmt`.