memory: add formatting utility for memory declarations
Info attributes are a property of a statement, which may be a
declaration. However, due to the weird formatting those FIRRTL people
came up with we actually need to format the info attribute as part of a
memory declaration. Which may make sense if your data model is a bunch
of java/scala classes inheriting from another anyway. However, for us
it's a bit of a bother, and the existing `DisplayIndented` impl for
`Memory` isn't sufficient any more.

The utility introduced with this change allows formatting a `Memory`
alongside an `Info` annotation. It will replace the existing
`DisplayIndented` impl for `Memory` itself, which will be removed in the
future.