memory: make memory parser parse info attribute
Info attributes are a property of a statement, which may be a
declaration. However, due to the weird formatting those FIRRTL people
came up with we actually need to parse the info attribute as part of a
memory declaration. Which may make sense if your data model is a bunch
of java/scala classes inheriting from another anyway. However, for us
it's a bit of a bother.

This change makes the `memory` parser parse a `Memory` with an info
attribute and return both so that in the future we can use both for
construction of a statement.