stmt: add utility for formatting an "optional name"
FIRRTL defines optional name attributes for halt and printf statements.
In both cases, they are separated from the statement with a colon (`:`),
which only appears if a name is set. Implementing this in the frmatting
implementation of `Statement` itself would lead to nasty clutter. Hence,
we do what we always do in such situations and introduce a formatting
utility just for those names.