stmt: add optional name attribute to halt and printf statements
FIRRTL defines optional name attributes for halt and printf statements.
We ignored them in the past, but ultimately we obviously need to support
them in some way.

We choose to represent them in the AST, meaning that `Kind::Stop` and
`Kind::Print` gain new attributes. For now, we neither print nor parse
them and always initialize them with a default value. However, future
changes will extend `Statement`'s `Display` impl as well as the `stmt`
parser.