circuit: make ModuleConsumer carry an info
After retrieving a circuit's "head", e.g. by parsing its first line, we
already have both the top module's name and the circuits info attribute.
Having to drag around both a `ModuleConsumer` and the info is not that
ergonomic. So we decided to integrate the latter into the former. Since
it's optional, anyway, users can still omit it if they want to.