error: add an error type for parsing
We plan on making parsing more ergonomic for the user, including a
"parse from file"-functionality. The `Result` such a function would
return would need to express both general IO-errors and actual parse
errors.

As we don't necessarily want to pass `nom` error types to the user (and
we'd run into lifetime issues if we did), we introduce a rather opaque
variant which just contains an error string.