circuit: make circuit parser return a ParseError
This change also eliminates an issue where a parse failure would
inevitably result in the emission of an obscuring error since we
obviously wouldn't find the top module either.

The parse test is updated accordingly, and as a result it no longer
detects cases where we don't parse the entire input. We don't deem this
a problem, however.