circuit: make ModuleConsumer process Results
We originally planned to have `ModuleConsumer` wrap an `Iterator` over
`Results` but decided otherwise. This choice was mainly motivated by
`nom::combinator::ParserIterator` yielding parsed items and delivering a
result on the `finish` call.

However, we plan on making `module::Modules` an `Iterator` which returns
errors early and does not depend on any finalizing calls. Such an
`Iterator` would return an `Result`.