module: make Modules an Iterator
`Modules` provides an interface of parsing `Module`s while managing the
associated namespace relevant for module instantiations. Previously, it
was driven by callers, which needed to supply both an input and an
indentation. Especially the latter would be carries alongside the
`Modules` instance in virtually any sane use-cases.

We plan on making `Modules` a part of the public API. However, we don't
want to expose the `Indentation` type nor any `nom` types, especially
not its `Error` types. We therefore integrate all of those into the
`Modules`, making it an `Iterator` which is created from the input
containing all module definitions and yielding `Module`s wraped in
`Result`s with our public `ParseError` as error type.