module: preserve original input in Modules
When converting a `nom::Err` to a `ParseError` we need to include the
original reference, which is used for computing line and column numbers
included in the error message. Previously, we would not preserve the
original input but update it after a successful parse. Thus, the line
number would always be relative to the beginning of a parse.