module: remove impl of Display for Kind
Up until now, `Kind` would be used only for denoting a `Module` as
either regular or external. However, we plan to make it a complete
representation of a "type" of `Module`, including the data associated to
(only) that kind (e.g. `Statement`s for regular modules). Ultimately
`Kind` will be used as a representation of those inside `Module`.

As a `Kind` will not longer reflect only the "type" of module, having a
`Display` impl formatting it solely as a keyword appears not particular
intuitive. We therefore decided to drop the impl and use `Kind::keyword`
directly in `Module` formatting code.