module: drop Copy impl for Kind
Up until now, `Kind` would be used only for denoting a `Module` as
either regular or external. However, we plan to make it a complete
representation of a "type" of `Module`, including the data associated to
(only) that kind (e.g. `Statement`s for regular modules). At least some
of that data will not implement `Copy`.