module: add fns for creating empty Kinds
Up until now, `Kind` would be used only for denoting a `Module` as
either regular or external. However, we plan to make it a complete
representation of a "type" of `Module`, including the data associated to
(only) that kind (e.g. `Statement`s for regular modules). Ultimately
`Kind` will be used as a representation of those inside `Module`.

Still, we will need to create "empty" variants of Kinds occasionally.
The functions introduced by this change will allow creating these values
without the need to specify every signle future data member.