module: make Kind::Regular store statements
Up until now, `Kind` would be used only for denoting a `Module` as
either regular or external. However, we plan to make it a complete
representation of a "type" of `Module`, including the data associated to
(only) that kind. Ultimately `Kind` will be used as a representation of
those inside `Module`.

This change makes `Kind` carry `Statement`s for `Regular` modules and
makes `Module` both initialize its `Statement` list from that data
member and include a copy in retrieved module `Kind`s.

For now, the latter results in a potentially costly deep copy. However,
we will migrate the data storage in `Module` from using an `Option` to a
`Kind` in the future, which will also allow us to return a reference
from `kind()`.