module: remove Module::statements_mut
In the past, `Kind` was used only for denoting a `Module` as either
regular or external. Recently however, we let it carry data associated
to a `Module` `Kind`. We are planning to use it inside `Module`,
replacing the current `stmt` member. Furthermore, we plan on allowing
mutable accesses to `Kind`-specific data only through mutable references
to an internal `Kind`.

We already migrated to a new construction sheme which doesn't rely on
`statements_mut`. Hence, we can now safely remove that function.