module: use Kind for storing Statements in a Module
In the past, `Kind` was used only for denoting a `Module` as either
regular or external. Recently however, we let it carry data associated
to a `Module` `Kind`, planning on using it for data storage inside
`Module`. This change finally relaces the `stmt` member, which was an
`Option`, with a `Kind`, which allows future extension more easily.

This change itself does not alter the `Module` interface itself in any
way. However, as past changes prepared wtih change future changes will
allow more efficient access of a `Module`s kind as well as mutating it.