module: extend quickcheck::Arbitrary impl of Kind
In the past, `Kind` was used only for denoting a `Module` as either
regular or external. Recently however, we let it carry data associated
to a `Module` `Kind` and started using it for data storage inside
`Module`. In short, we transfered some responsibility from `Module` to
`Kind`.

However, up until now `Module` would still generate and shrink data
stored in a `Kind` on its own. We obviously want to transfer the
responsbility for this over to `Kind`, too.

As a preparation, we make `Kind`s `Arbitrary` impl both generate a
statement list for the `Regular` variant and shrink it, so that in the
future, the impl for `Module` can rely on it.