module: make Arbitrary impl for Module rely on impl for Kind
In the past, `Kind` was used only for denoting a `Module` as either
regular or external. Recently however, we let it carry data associated
to a `Module` `Kind` and started using it for data storage inside
`Module`. In short, we transfered some responsibility from `Module` to
`Kind`.

However, up until now `Module` would still generate and shrink data
stored in a `Kind` on its own. However, thanks to recent changes, we can
not rely on `Kind` to generate meaningful data and shrink it, too. As a
result, the way we generate and shrink `Port`s changes, and we don't
enforce a maximum number of ports for regular modules any more but rely
on the statement list, which is used as a source of ports, being
bounded.