module: impl Display for ParamValue
Unfortunately, our control over floating point formatting is limited. We
can't enforce an `f64` to be formatted with a point without providing a
precision. We don't want to enforce an absurdly high precision, neither
do we want to calculate the precision on our own or write our own
formatter.

So we just give up and accept that an `f64` may be formatted in a way
users (and parsers) may confuse it with an integer.