parsers: make le consume comments
Up until now we ignored the possibility of FIRRTL sources containing
comments. They obviously can. As we are not interested in introducing
them into the AST, we want to ignore them, and we want to make this as
automatic as possible, i.e. not introduce yet another microparser
everywhere a comment may appear.

As in many other languages, comments in FIRRTL go from some indicator to
the end of a line. Thus, we can get rid of any comments by simply
considering them part of a line ending.