memory: add kind of simple memory
FIRRTL apparently knows not only `mem`, but also `cmem` and `smem`,
which appear to be simpler types of memory. As they are quite similar
apart from the fact that apparently only an `smem` can have a defined
read-under-write behaviour, we decided to throw both together in a
single type and use this `enum` for differentiating between the two.