memory: move definition of ReadUnderWrite to submodule for common types
Currently, the `memory` module only hosts `Memory`, which represents a
FIRRTL `mem`, and related types. However, we'll also introduce types
representing `cmem` and `smem` constructs in the future. As they are
also memories, and one of them includes a `ReadUnderWrite`, we'll want
to make them part of `memory`. As a result, the module will grow
significantly. We therefore decided to split it into several submodules.

As `ReadUnderWrite` will be used in multiple kinds of memory, we decided
to move it into a submodule for common types.