memory: add type representing a port directory
Currently, the `memory` module only hosts `Memory`, which represents a
FIRRTL `mem`, and related types. However, we'll also introduce types
representing `cmem` and `smem` constructs in the future.

All memories also have ports of some kind, which also have a direction.
We therefore decided to re-use the concept of a `PortKind`. However, we
also find the name will not fit the thing it expresses too well.
Instead, we'll go with `PortDir`.

Rather than moving `PortKind` and renaming it at the same time, we
decided to introduce a more or less identical type under the new name
and have it replace the old one.