stmt: sort references by name for parse_entities test
When writing the test initially, we were a bit lazy. We'd collect
referenced entities in a `Vec` and collect their names into a set in
order to check uniqueness. We wouldn't even use the benefits of a hash
map for lookup, resulting in `O(n^2)` complexity for the overall test.

Just sorting `Vec` items by name allows both efficient checks for
uniqueness without allocating additional memory and use a banary search
for the lookup.

We already introduced this enhancement for the `parse_stmt` test with

        2063a994b34fc86f66b2b9c137a4c2b8324b59a5

but overlooked `parse_entity`.