stmt: make entity known only after its dependencies in stmt_with_decls
In `stmt_with_decls`, the `entity` map is used for keeping track of
entities which are already declared by previous statements. However, we
previously would add declarations first to this map and then extend the
statement. In theory, this could lead to the generation of statements
with circular dependencies.

This change moves the addition of a declaration after the extension and
also makes sure its name is indeed unique. This change also allows us to
consider only a single declaration statement at a given point in time
rather than pre-iterating over all declarations in a statement.