stmt: make SubContext add items to parent context when dropped
In the past, we operated under the assumption that declarations were
visible only within their "scope", e.g. only within a `when` or `else`
branch. Apparently, this is not the case. Rather, declarations appear to
be visible also in statements following the branching statement. Thus,
we need to add items we got inside a scope to the parent context when
we're done.