named: add a trait for named items
We have quite a few types which have a "name" property which is either
exposed through a fn `name`, which exposes the name as either an
`&Arc<str>` or a `&str`, or the `Reference` trait, which exposes the
name as a `&str`. For end-users, `&str` will usually be more convenient.
However, we do want to share as much constant data as possible, which
motivated storing names as `Arc<str>`s and exposing those for some
types.

This trait will allow accessing items' names both as the type they are
stored in and as a `&str`. We choose to make the storage type generic to
keep our tests simple and in order to accommodate for future use-cases.

Future changes will implement this trait for all relevant types, in
particular, all `Reference`s. The goal will be to make `Reference`
depend on `Named` and remove its `name` function.