types: add types to use for number of bits
We currently express "number of bits" via `u16` and `i16` in many
places. This is not ideal: if we were to change the width expressible in
a `BitWidth`, it would create fallout in many places.

This change introduces new types to use for expressing "number of bits".
Future changes will migrate all instances we recently used `u16` and
`i16` directly.