types: add function for retrieving components of a Type::Vector
`Type` already featured a function `vector_base` which allowed
retrieving a vectors base type in a convenient way, i.e. wrapped in an
`Option`, saving the need for matching a `Type` against a `Type::Vector`
by the caller. However, we neglected possible use-cases for which a user
might want to retreive the width of a vector, or both base type and
width.