memory: simplify shrinking of mem::Memory
In many cases, implementations of `quickcheck::Arbitrary::shrink` need
to consider multiple inner states which may be shrunk independently from
another. In the past, we often built shrinkers for each of those states
and chained them together.

Shrinking a tuple of those independent states serves the same purpose or
rather delegates proper shrinking of a combination to `quickcheck`. This
approach is both more concise and arguably more expressive. We therefore
decided to migrate to that approach for the sake of maintainability.