types: impl PartialEq<GroundType> for Type
... for convenience.

We suspect that comparing a `Type` against some predefined `GroundType`
will be rather common. The impl introduced by this change will eliminate
boiler plate necessary in such situations, namely a conversion from
`GroundType` to `Type`.

This change may break user code relying on `Into::into` for such a
conversion, as an expression `a == b.into()` with `a: Type` and
`b: GroundTypoe` will now be ambiguous regarding the `PartialEq` and
`Into` impl to use.